requirements.txt
MetaTrader5>=5.0.45
stable-baselines3>=2.1.0
gymnasium>=0.29.0
numpy>=1.24.0
pandas>=2.0.0
pandas-ta>=0.3.25b
python-dotenv>=1.0.0
shimmy>=1.3.0
pyarrow>=14.0.0
